==> f03_wes_wecare_R/s08_crude_counts/s03_allele_counts_bc11.Rmd <==
---
title: "WES: BC11 crude allele counts, WECARE only"
author: "Alexey Larionov"
date: "22 Apr 2022"
output:
  html_document:
    toc: true
    toc_float: true
editor_options:
  chunk_output_type: console
---

# Summary

- Calculate crude counts: totals, and separately for cases and controls  
- Fisher test for the crude counts  

**Aggregated** AF is defined as **total** AC over **mean** AN in the studied variants  

<style>
pre{
  overflow-x: auto;
}
pre code{
  word-wrap: normal;
  white-space: pre;
}
</style>

```{r echo=F}
options(width=999)
```

# Start section

```{r}

# Time
Sys.time()

# Memory
gc()

# Clean up
rm(list=ls())
graphics.off()

# Options
options(stringsAsFactors = F)

# Working folders
base_folder <- "/Users/alexey/Documents"
project_folder <- file.path(base_folder,"wecare","reanalysis_2022","wes","s01_wes_wecare_only")
scripts_folder <- file.path(project_folder,"scripts","s08_crude_counts")
setwd(scripts_folder)
data_folder <- file.path(project_folder,"data","s08_crude_counts")

# Libraries
library(dplyr)

```

# Load data

```{r}

load(file.path(data_folder,"s02_select_bc11_pv.RData"))

# Update working folders
base_folder <- "/Users/alexey/Documents"
project_folder <- file.path(base_folder,"wecare","reanalysis_2022","wes","s01_wes_wecare_only")
scripts_folder <- file.path(project_folder,"scripts","s08_crude_counts")
setwd(scripts_folder)
data_folder <- file.path(project_folder,"data","s08_crude_counts")

```

# Check data

```{r}

ls()

# Check objects dimensions
dim(bc11_variants.df)
dim(bc11_fiv.df)

dim(genotypes.mx)
dim(phenotypes.df)
dim(variants.df)

# Check that the tables and matrices are in sync
sum(phenotypes.df$wes_id != colnames(genotypes.mx))
sum(variants.df$ID != rownames(genotypes.mx))

table(phenotypes.df$cc, useNA = "always")

# Check list of genes
bc11_genes

# Check mean age
mean(phenotypes.df$age_dx)
sd(phenotypes.df$age_dx)

# Check distribution of RT
table(phenotypes.df$cc)
table(phenotypes.df$br_xray)
x <- table(phenotypes.df$cc, phenotypes.df$br_xray)
c("UBC (172)","CBC (162)") -> rownames(x)
c("Rx- (161)","Rx+ (173)") -> colnames(x)
x
fisher.test(x)
rm(x)

```

# Aggregated crude counts

## Total

```{r}

total_ac <- sum(bc11_fiv.df$total_ac) 

total_an <- sum(bc11_fiv.df$total_an)
mean_an <- total_an / nrow(bc11_fiv.df)

mean_af <- total_ac / total_an
aggregated_af <- total_ac / mean_an

mean_af
aggregated_af

rm(total_ac,total_an,mean_an,mean_af,aggregated_af)

```

## Controls

```{r}

controls_ac <- sum(bc11_fiv.df$controls_ac) 

controls_total_an <- sum(bc11_fiv.df$controls_an)
controls_mean_an <- controls_total_an / nrow(bc11_fiv.df)

controls_mean_af <- controls_ac / controls_total_an
controls_aggregated_af <- controls_ac / controls_mean_an

rm(controls_mean_an)

```

## Cases

```{r}

cases_ac <- sum(bc11_fiv.df$cases_ac)

cases_total_an <- sum(bc11_fiv.df$cases_an)
cases_mean_an <- cases_total_an / nrow(bc11_fiv.df)

cases_mean_af <- cases_ac / cases_total_an
cases_aggregated_af <- cases_ac / cases_mean_an

rm(cases_mean_an)

```

## Fisher test

```{r}

fisher_matrix <- matrix(c(cases_ac, cases_total_an - cases_ac,
                          controls_ac, controls_total_an - controls_ac),
                          nrow=2, byrow=T)

colnames(fisher_matrix) <- c("Alt","Ref")
rownames(fisher_matrix) <- c("Cases","Controls") # CBC, UBC
fisher_matrix 

ft <- fisher.test(fisher_matrix, conf.int=T)

ft

rm(cases_ac,cases_total_an,controls_ac,controls_total_an,fisher_matrix)

```

## Barplot

### Mean AF

```{r}

# Prepare data for plot
afs <- c(controls_mean_af,cases_mean_af)
names(afs) <- c("UBC","CBC")

cc_counts <- table(phenotypes.df$cc)

# Barplot (keep mid-positions of bars)
mp <- barplot(afs, ylim=c(0,max(afs)+0.0005), 
              main=paste("BC11: WES WECARE-only\n",
                cc_counts[1],"controls vs ",cc_counts[2],
                " cases\nCrude counts Fisher p=",round(ft$p.value,3)),
              ylab="Mean AF in selected variants")

# Add labels to plot
text(mp,afs+0.0002,round(afs,4))

# Clean-up
rm(controls_mean_af, cases_mean_af, afs, mp)

```

### Aggregated AF

```{r}

# Prepare data for plot
afs <- c(controls_aggregated_af,cases_aggregated_af)
names(afs) <- c("UBC","CBC")

# Barplot (keep mid-positions of bars)
mp <- barplot(afs, ylim=c(0,max(afs)+0.005), 
              main=paste("BC11 Aggregated AF: WES WECARE-only\n",
                cc_counts[1],"controls vs ",cc_counts[2],
                " cases\nCrude counts Fisher p=",round(ft$p.value,3)),
              ylab="Aggregated AF")

# Add labels to plot
text(mp,afs+0.003,round(afs,3))

# Barplot with a simplified tiple
mp <- barplot(afs, ylim=c(0,max(afs)+0.005), 
              main="Aggregated AF",
              ylab=NA)

# Add labels to plot
text(mp,afs+0.003,round(afs,3))

# Clean-up
rm(controls_aggregated_af, cases_aggregated_af, afs, mp, cc_counts, ft)

```

# Check data

```{r}

ls()

# Check objects dimentions
dim(bc11_variants.df)
dim(bc11_fiv.df)

dim(genotypes.mx)
dim(phenotypes.df)
dim(variants.df)

# Check that the tables and matrices are in sync
sum(phenotypes.df$wes_id != colnames(genotypes.mx))
sum(variants.df$ID != rownames(genotypes.mx))

table(phenotypes.df$cc, useNA = "always")

```

# Save results

```{r}

save.image(file.path(data_folder,"s03_allele_counts_bc11.RData"))

```

# Final section

```{r}

sessionInfo()
Sys.time()
gc()

```
